+++ test/data.jl
@@ -1,5 +1,7 @@
 # Tests of data checking and transformation code
 
+TESTLOG = getlogger(PowerModels)
+
 @testset "test data summary" begin
 
     @testset "5-bus summary from dict" begin
@@ -430,3 +432,53 @@ end
     @test isapprox(g.values, [-1.0 0.5 0; 0.75 -0.25 0; 0 0 0])
     @test isapprox(b.values, [1.0 -0.5 0; -0.75 0.25 0; 0 0 0])
 end
+
+
+
+@testset "test branch flow computations" begin
+     @testset "5-bus ac polar flow" begin
+        data = PowerModels.parse_file("../test/data/matpower/case5.m")
+        result = run_opf(data, ACPPowerModel, ipopt_solver; setting = Dict("output" => Dict("branch_flows" => true)))
+        PowerModels.update_data(data, result["solution"])
+
+        ac_flows = PowerModels.calc_branch_flow_ac(data)
+
+        for (i,branch) in data["branch"]
+            branch_flow = ac_flows["branch"][i]
+            for k in ["pf","pt","qf","qt"]
+                @test isapprox(branch[k], branch_flow[k]; atol=1e-6)
+            end
+        end
+    end
+
+    @testset "5-bus ac rect flow" begin
+        data = PowerModels.parse_file("../test/data/matpower/case5.m")
+        result = run_opf(data, ACRPowerModel, ipopt_solver; setting = Dict("output" => Dict("branch_flows" => true)))
+        PowerModels.update_data(data, result["solution"])
+
+        ac_flows = PowerModels.calc_branch_flow_ac(data)
+
+        for (i,branch) in data["branch"]
+            branch_flow = ac_flows["branch"][i]
+            for k in ["pf","pt","qf","qt"]
+                @test isapprox(branch[k], branch_flow[k]; atol=1e-6)
+            end
+        end
+    end
+
+    @testset "5-bus dc flow" begin
+        data = PowerModels.parse_file("../test/data/matpower/case5.m")
+        result = run_opf(data, DCPPowerModel, ipopt_solver; setting = Dict("output" => Dict("branch_flows" => true)))
+        PowerModels.update_data(data, result["solution"])
+
+        dc_flows = PowerModels.calc_branch_flow_dc(data)
+
+        for (i,branch) in data["branch"]
+            branch_flow = dc_flows["branch"][i]
+            for k in ["pf","pt"]
+                @test isapprox(branch[k], branch_flow[k]; atol=1e-6)
+            end
+        end
+    end
+
+end